Clean up notebooks and make code exit gracefully when exception occurs.
