+++ src/astrohack/extract_holog.py
@@ -130,287 +130,293 @@ def extract_holog(
             }
 
     """
-    extract_holog_params = locals()
-
-    logger = _get_astrohack_logger()
-
-    function_name = inspect.stack()[CURRENT_FUNCTION].function
-
-    ######### Parameter Checking #########
-    extract_holog_params = _check_extract_holog_params(function_name=function_name, extract_holog_params=extract_holog_params)
-    
-
-    _check_if_file_exists(extract_holog_params['ms_name'])
-    _check_if_file_will_be_overwritten(extract_holog_params['holog_name'], extract_holog_params['overwrite'])
-
-    if holog_name==None:
-        
-        logger.debug('[{caller}]: File {file} does not exists. Extracting ...'.format(caller=function_name, file=holog_name))
-            
-        from astrohack._utils._tools import _remove_suffix
-
-        holog_name = _remove_suffix(ms_name, '.ms') + '.holog.zarr'
-        extract_holog_params['holog_name'] = holog_name
-            
-        logger.debug('[{caller}]: Extracting holog to {output}'.format(caller=function_name, output=holog_name))
-          
     try:
-        pnt_dict = _load_point_file(extract_holog_params['point_name'])
-    except:
-        logger.warning(f'[{function_name}]: Could not find {extract_holog_params["point_name"]}, creating point new 'f'point.zarr.')        
+        extract_holog_params = locals()
 
-    ######## Get Spectral Windows ########
-    ctb = ctables.table(
-        os.path.join(extract_holog_params['ms_name'], "DATA_DESCRIPTION"),
-        readonly=True,
-        lockoptions={"option": "usernoread"},
-        ack=False,
-    )
-    ddi_spw = ctb.getcol("SPECTRAL_WINDOW_ID")
-    ddpol_indexol = ctb.getcol("POLARIZATION_ID")
-    ms_ddi = np.arange(len(ddi_spw))
-    ctb.close()
+        logger = _get_astrohack_logger()
 
-    ######## Get Antenna IDs and Names ########
-    ctb = ctables.table(
-        os.path.join(extract_holog_params['ms_name'], "ANTENNA"),
-        readonly=True,
-        lockoptions={"option": "usernoread"},
-        ack=False,
-    )
+        function_name = inspect.stack()[CURRENT_FUNCTION].function
 
-    ant_names = np.array(ctb.getcol("NAME"))
-    ant_id = np.arange(len(ant_names))
-    ant_pos = ctb.getcol("POSITION")
-
-    ctb.close()
+        ######### Parameter Checking #########
+        extract_holog_params = _check_extract_holog_params(function_name=function_name, extract_holog_params=extract_holog_params)
     
-    ######## Get Antenna IDs that are in the main table########
-    ctb = ctables.table(
-        extract_holog_params['ms_name'],
-        readonly=True,
-        lockoptions={"option": "usernoread"},
-        ack=False,
-    )
 
-    ant1 = np.unique(ctb.getcol("ANTENNA1"))
-    ant2 = np.unique(ctb.getcol("ANTENNA2"))
-    ant_id_main = np.unique(np.append(ant1,ant2))
-    
-    ant_names_main = ant_names[ant_id_main]
-    ctb.close()
-    
-    # Create holog_obs_dict or modify user supplied holog_obs_dict.
-    ddi = extract_holog_params['ddi']
+        _check_if_file_exists(extract_holog_params['ms_name'])
+        _check_if_file_will_be_overwritten(extract_holog_params['holog_name'], extract_holog_params['overwrite'])
 
-    if holog_obs_dict is None: #Automatically create holog_obs_dict
-        from astrohack._utils._extract_holog import _create_holog_obs_dict
-        holog_obs_dict = _create_holog_obs_dict(
-            pnt_dict, 
-            extract_holog_params['baseline_average_distance'],
-            extract_holog_params['baseline_average_nearest'], 
-            ant_names, 
-            ant_pos,
-            ant_names_main
+        if holog_name==None:
+            
+            logger.debug('[{caller}]: File {file} does not exists. Extracting ...'.format(caller=function_name, file=holog_name))
+            
+            from astrohack._utils._tools import _remove_suffix
+
+            holog_name = _remove_suffix(ms_name, '.ms') + '.holog.zarr'
+            extract_holog_params['holog_name'] = holog_name
+            
+            logger.debug('[{caller}]: Extracting holog to {output}'.format(caller=function_name, output=holog_name))
+          
+        try:
+            pnt_dict = _load_point_file(extract_holog_params['point_name'])
+        except:
+            logger.warning(f'[{function_name}]: Could not find {extract_holog_params["point_name"]}, creating point new 'f'point.zarr.')        
+
+        ######## Get Spectral Windows ########
+        ctb = ctables.table(
+            os.path.join(extract_holog_params['ms_name'], "DATA_DESCRIPTION"),
+            readonly=True,
+            lockoptions={"option": "usernoread"},
+            ack=False,
         )
         
-        #From the generated holog_obs_dict subselect user supplied ddis.
-        if ddi != 'all':
-            holog_obs_dict_keys = list(holog_obs_dict.keys())
-            for ddi_key in holog_obs_dict_keys:
-                if 'ddi' in ddi_key:
-                    ddi_id = int(ddi_key.replace('ddi_',''))
-                    if ddi_id not in ddi:
-                        del holog_obs_dict[ddi_key]
+        ddi_spw = ctb.getcol("SPECTRAL_WINDOW_ID")
+        ddpol_indexol = ctb.getcol("POLARIZATION_ID")
+        ms_ddi = np.arange(len(ddi_spw))
+        ctb.close()
 
-    else:
-        #If a user defines a holog_obs_dict it needs to be duplicated for each ddi.
-        holog_obs_dict_with_ddi = {}
-        if ddi == 'all':
-            for ddi_id in ms_ddi:
-                holog_obs_dict_with_ddi['ddi_' + str(ddi_id)] = holog_obs_dict
-        else:
-            for ddi_id in ddi:
-                holog_obs_dict_with_ddi['ddi_' + str(ddi_id)] = holog_obs_dict
-        
-        holog_obs_dict = holog_obs_dict_with_ddi
-            
-    logger.info(f"[{function_name}]: holog_obs_dict: \n%s", pformat(list(holog_obs_dict.values())[0], indent=2, width=2))
+        ######## Get Antenna IDs and Names ########
+        ctb = ctables.table(
+            os.path.join(extract_holog_params['ms_name'], "ANTENNA"),
+            readonly=True,
+            lockoptions={"option": "usernoread"},
+            ack=False,
+        )
 
+        ant_names = np.array(ctb.getcol("NAME"))
+        ant_id = np.arange(len(ant_names))
+        ant_pos = ctb.getcol("POSITION")
 
-    outfile_obj = copy.deepcopy(holog_obs_dict)
-
-    _jsonify(outfile_obj)
-
-    with open(".holog_obs_dict.json", "w") as outfile:
-        json.dump(outfile_obj, outfile)
-        
-    ######## Get Scan and Subscan IDs ########
-    # SDM Tables Short Description (https://drive.google.com/file/d/16a3g0GQxgcO7N_ZabfdtexQ8r2jRbYIS/view)
-    # 2.54 ScanIntent (p. 150)
-    # MAP ANTENNA SURFACE : Holography calibration scan
-
-    # 2.61 SubscanIntent (p. 152)
-    # MIXED : Pointing measurement, some antennas are on-source, some off-source
-    # REFERENCE : reference measurement (used for boresight in holography).
-    # Undefined : ?
-
-    ctb = ctables.table(
-        os.path.join(extract_holog_params['ms_name'], "STATE"),
-        readonly=True,
-        lockoptions={"option": "usernoread"},
-        ack=False,
-    )
-
-    # scan intent (with subscan intent) is stored in the OBS_MODE column of the STATE subtable.
-    obs_modes = ctb.getcol("OBS_MODE")
-    ctb.close()
-
-    scan_intent = "MAP_ANTENNA_SURFACE"
-    state_ids = []
-
-    for i, mode in enumerate(obs_modes):
-        if (scan_intent in mode) and ('REFERENCE' not in mode):
-            state_ids.append(i)
-
-    spw_ctb = ctables.table(
-        os.path.join(extract_holog_params['ms_name'], "SPECTRAL_WINDOW"),
-        readonly=True,
-        lockoptions={"option": "usernoread"},
-        ack=False,
-    )
-    pol_ctb = ctables.table(
-        os.path.join(extract_holog_params['ms_name'], "POLARIZATION"),
-        readonly=True,
-        lockoptions={"option": "usernoread"},
-        ack=False,
-    )
-
-    obs_ctb = ctables.table(
-        os.path.join(extract_holog_params['ms_name'], "OBSERVATION"),
-        readonly=True,
-        lockoptions={"option": "usernoread"},
-        ack=False,
-    )
+        ctb.close()
     
-    telescope_name = obs_ctb.getcol("TELESCOPE_NAME")[0]
-    start_time_unix = obs_ctb.getcol('TIME_RANGE')[0][0] - 3506716800.0
-    time = Time(start_time_unix, format='unix').jyear
+        ######## Get Antenna IDs that are in the main table########
+        ctb = ctables.table(
+            extract_holog_params['ms_name'],
+            readonly=True,
+            lockoptions={"option": "usernoread"},
+            ack=False,
+        )
 
-    # If we have an EVLA run from before 2023 the pointing table needs to be fixed.
-    if telescope_name == "EVLA" and time < 2023:
-        # Convert from casa epoch to unix time
-        his_ctb = ctables.table(
-            os.path.join(extract_holog_params['ms_name'], "HISTORY"),
+        ant1 = np.unique(ctb.getcol("ANTENNA1"))
+        ant2 = np.unique(ctb.getcol("ANTENNA2"))
+        ant_id_main = np.unique(np.append(ant1,ant2))
+    
+        ant_names_main = ant_names[ant_id_main]
+        ctb.close()
+    
+        # Create holog_obs_dict or modify user supplied holog_obs_dict.
+        ddi = extract_holog_params['ddi']
+
+        if holog_obs_dict is None: #Automatically create holog_obs_dict
+            from astrohack._utils._extract_holog import _create_holog_obs_dict
+            holog_obs_dict = _create_holog_obs_dict(
+                pnt_dict, 
+                extract_holog_params['baseline_average_distance'],
+                extract_holog_params['baseline_average_nearest'], 
+                ant_names, 
+                ant_pos,
+                ant_names_main
+            )
+        
+            #From the generated holog_obs_dict subselect user supplied ddis.
+            if ddi != 'all':
+                holog_obs_dict_keys = list(holog_obs_dict.keys())
+                for ddi_key in holog_obs_dict_keys:
+                    if 'ddi' in ddi_key:
+                        ddi_id = int(ddi_key.replace('ddi_',''))
+                        if ddi_id not in ddi:
+                            del holog_obs_dict[ddi_key]
+
+        else:
+            #If a user defines a holog_obs_dict it needs to be duplicated for each ddi.
+            holog_obs_dict_with_ddi = {}
+            if ddi == 'all':
+                for ddi_id in ms_ddi:
+                    holog_obs_dict_with_ddi['ddi_' + str(ddi_id)] = holog_obs_dict
+            else:
+                for ddi_id in ddi:
+                    holog_obs_dict_with_ddi['ddi_' + str(ddi_id)] = holog_obs_dict
+        
+            holog_obs_dict = holog_obs_dict_with_ddi
+            
+        logger.info(f"[{function_name}]: holog_obs_dict: \n%s", pformat(list(holog_obs_dict.values())[0], indent=2, width=2))
+
+
+        outfile_obj = copy.deepcopy(holog_obs_dict)
+
+        _jsonify(outfile_obj)
+
+        with open(".holog_obs_dict.json", "w") as outfile:
+            json.dump(outfile_obj, outfile)
+        
+        ######## Get Scan and Subscan IDs ########
+        # SDM Tables Short Description (https://drive.google.com/file/d/16a3g0GQxgcO7N_ZabfdtexQ8r2jRbYIS/view)
+        # 2.54 ScanIntent (p. 150)
+        # MAP ANTENNA SURFACE : Holography calibration scan
+
+        # 2.61 SubscanIntent (p. 152)
+        # MIXED : Pointing measurement, some antennas are on-source, some off-source
+        # REFERENCE : reference measurement (used for boresight in holography).
+        # Undefined : ?
+
+        ctb = ctables.table(
+            os.path.join(extract_holog_params['ms_name'], "STATE"),
+            readonly=True,
+            lockoptions={"option": "usernoread"},
+            ack=False,
+        )
+
+        # scan intent (with subscan intent) is stored in the OBS_MODE column of the STATE subtable.
+        obs_modes = ctb.getcol("OBS_MODE")
+        ctb.close()
+
+        scan_intent = "MAP_ANTENNA_SURFACE"
+        state_ids = []
+
+        for i, mode in enumerate(obs_modes):
+            if (scan_intent in mode) and ('REFERENCE' not in mode):
+                state_ids.append(i)
+
+        spw_ctb = ctables.table(
+            os.path.join(extract_holog_params['ms_name'], "SPECTRAL_WINDOW"),
+            readonly=True,
+            lockoptions={"option": "usernoread"},
+            ack=False,
+        )
+
+        pol_ctb = ctables.table(
+            os.path.join(extract_holog_params['ms_name'], "POLARIZATION"),
+            readonly=True,
+            lockoptions={"option": "usernoread"},
+            ack=False,
+        )
+
+        obs_ctb = ctables.table(
+            os.path.join(extract_holog_params['ms_name'], "OBSERVATION"),
             readonly=True,
             lockoptions={"option": "usernoread"},
             ack=False,
         )
     
-        assert ("pnt_tbl:fixed" in his_ctb.getcol("MESSAGE")), \
-            "Pointing table not corrected, users should apply function astrohack.dio.fix_pointing_table() to " \
-            "remedy this."
+        telescope_name = obs_ctb.getcol("TELESCOPE_NAME")[0]
+        start_time_unix = obs_ctb.getcol('TIME_RANGE')[0][0] - 3506716800.0
+        time = Time(start_time_unix, format='unix').jyear
+
+        # If we have an EVLA run from before 2023 the pointing table needs to be fixed.
+        if telescope_name == "EVLA" and time < 2023:
+            # Convert from casa epoch to unix time
+            his_ctb = ctables.table(
+                os.path.join(extract_holog_params['ms_name'], "HISTORY"),
+                readonly=True,
+                lockoptions={"option": "usernoread"},
+                ack=False,
+            )
+    
+            assert ("pnt_tbl:fixed" in his_ctb.getcol("MESSAGE")), \
+                "Pointing table not corrected, users should apply function astrohack.dio.fix_pointing_table() to " \
+                "remedy this."
         
         his_ctb.close()
 
-    count = 0
-    delayed_list = []
-    for ddi_name in holog_obs_dict.keys():
-        ddi = int(ddi_name.replace('ddi_',''))
-        spw_setup_id = ddi_spw[ddi]
-        pol_setup_id = ddpol_indexol[ddi]
+        count = 0
+        delayed_list = []
+        for ddi_name in holog_obs_dict.keys():
+            ddi = int(ddi_name.replace('ddi_',''))
+            spw_setup_id = ddi_spw[ddi]
+            pol_setup_id = ddpol_indexol[ddi]
         
-        extract_holog_params["ddi"] = ddi
-        extract_holog_params["chan_setup"] = {}
-        extract_holog_params["pol_setup"] = {}
-        extract_holog_params["chan_setup"]["chan_freq"] = spw_ctb.getcol("CHAN_FREQ", startrow=spw_setup_id, nrow=1)[0, :]
-        extract_holog_params["chan_setup"]["chan_width"] = spw_ctb.getcol("CHAN_WIDTH", startrow=spw_setup_id, nrow=1)[0, :]
-        extract_holog_params["chan_setup"]["eff_bw"] = spw_ctb.getcol("EFFECTIVE_BW", startrow=spw_setup_id, nrow=1)[0, :]
-        extract_holog_params["chan_setup"]["ref_freq"] = spw_ctb.getcol("REF_FREQUENCY", startrow=spw_setup_id, nrow=1)[0]
-        extract_holog_params["chan_setup"]["total_bw"] = spw_ctb.getcol("TOTAL_BANDWIDTH", startrow=spw_setup_id, nrow=1)[0]
-        extract_holog_params["pol_setup"]["pol"] = pol_str[pol_ctb.getcol("CORR_TYPE", startrow=pol_setup_id, nrow=1)[0, :]]
+            extract_holog_params["ddi"] = ddi
+            extract_holog_params["chan_setup"] = {}
+            extract_holog_params["pol_setup"] = {}
+            extract_holog_params["chan_setup"]["chan_freq"] = spw_ctb.getcol("CHAN_FREQ", startrow=spw_setup_id, nrow=1)[0, :]
+            extract_holog_params["chan_setup"]["chan_width"] = spw_ctb.getcol("CHAN_WIDTH", startrow=spw_setup_id, nrow=1)[0, :]
+            extract_holog_params["chan_setup"]["eff_bw"] = spw_ctb.getcol("EFFECTIVE_BW", startrow=spw_setup_id, nrow=1)[0, :]
+            extract_holog_params["chan_setup"]["ref_freq"] = spw_ctb.getcol("REF_FREQUENCY", startrow=spw_setup_id, nrow=1)[0]
+            extract_holog_params["chan_setup"]["total_bw"] = spw_ctb.getcol("TOTAL_BANDWIDTH", startrow=spw_setup_id, nrow=1)[0]
+            extract_holog_params["pol_setup"]["pol"] = pol_str[pol_ctb.getcol("CORR_TYPE", startrow=pol_setup_id, nrow=1)[0, :]]
         
-        extract_holog_params["telescope_name"] = obs_ctb.getcol("TELESCOPE_NAME")[0]
+            extract_holog_params["telescope_name"] = obs_ctb.getcol("TELESCOPE_NAME")[0]
 
-        for holog_map_key in holog_obs_dict[ddi_name].keys(): #loop over all beam_scan_ids, a beam_scan_id can conist out of more than one scan in an ms (this is the case for the VLA pointed mosiacs).
+            for holog_map_key in holog_obs_dict[ddi_name].keys(): #loop over all beam_scan_ids, a beam_scan_id can conist out of more than one scan in an ms (this is the case for the VLA pointed mosiacs).
 
-            if 'map' in holog_map_key:
-                scans = holog_obs_dict[ddi_name][holog_map_key]["scans"]
-                logger.info(f"[{function_name}]: Processing ddi: {ddi}, scans: {scans}")
+                if 'map' in holog_map_key:
+                    scans = holog_obs_dict[ddi_name][holog_map_key]["scans"]
+                    logger.info(f"[{function_name}]: Processing ddi: {ddi}, scans: {scans}")
                 
-                if len(list(holog_obs_dict[ddi_name][holog_map_key]['ant'].keys())) != 0:
-                    map_ant_list = []
-                    ref_ant_per_map_ant_list = []
+                    if len(list(holog_obs_dict[ddi_name][holog_map_key]['ant'].keys())) != 0:
+                        map_ant_list = []
+                        ref_ant_per_map_ant_list = []
                     
-                    map_ant_name_list = []
-                    ref_ant_per_map_ant_name_list = []
-                    for map_ant_str in holog_obs_dict[ddi_name][holog_map_key]['ant'].keys():
-                        ref_ant_ids = np.array(_convert_ant_name_to_id(ant_names, list(holog_obs_dict[ddi_name][holog_map_key]['ant'][map_ant_str])))
+                        map_ant_name_list = []
+                        ref_ant_per_map_ant_name_list = []
+                        for map_ant_str in holog_obs_dict[ddi_name][holog_map_key]['ant'].keys():
+                            ref_ant_ids = np.array(_convert_ant_name_to_id(ant_names, list(holog_obs_dict[ddi_name][holog_map_key]['ant'][map_ant_str])))
                         
-                        map_ant_id = _convert_ant_name_to_id(ant_names,map_ant_str)[0]
+                            map_ant_id = _convert_ant_name_to_id(ant_names,map_ant_str)[0]
 
-                        ref_ant_per_map_ant_list.append(ref_ant_ids)
-                        map_ant_list.append(map_ant_id)
+                            ref_ant_per_map_ant_list.append(ref_ant_ids)
+                            map_ant_list.append(map_ant_id)
                         
-                        ref_ant_per_map_ant_name_list.append(list(holog_obs_dict[ddi_name][holog_map_key]['ant'][map_ant_str]))
-                        map_ant_name_list.append(map_ant_str)
+                            ref_ant_per_map_ant_name_list.append(list(holog_obs_dict[ddi_name][holog_map_key]['ant'][map_ant_str]))
+                            map_ant_name_list.append(map_ant_str)
 
-                    extract_holog_params["ref_ant_per_map_ant_tuple"] = tuple(ref_ant_per_map_ant_list)
-                    extract_holog_params["map_ant_tuple"] = tuple(map_ant_list)
+                        extract_holog_params["ref_ant_per_map_ant_tuple"] = tuple(ref_ant_per_map_ant_list)
+                        extract_holog_params["map_ant_tuple"] = tuple(map_ant_list)
                     
-                    extract_holog_params["ref_ant_per_map_ant_name_tuple"] = tuple(ref_ant_per_map_ant_name_list)
-                    extract_holog_params["map_ant_name_tuple"] = tuple(map_ant_name_list)
+                        extract_holog_params["ref_ant_per_map_ant_name_tuple"] = tuple(ref_ant_per_map_ant_name_list)
+                        extract_holog_params["map_ant_name_tuple"] = tuple(map_ant_name_list)
                     
-                    extract_holog_params["scans"] = scans
-                    extract_holog_params["sel_state_ids"] = state_ids
-                    extract_holog_params["holog_map_key"] = holog_map_key
-                    extract_holog_params["ant_names"] = ant_names
+                        extract_holog_params["scans"] = scans
+                        extract_holog_params["sel_state_ids"] = state_ids
+                        extract_holog_params["holog_map_key"] = holog_map_key
+                        extract_holog_params["ant_names"] = ant_names
                     
-                    if parallel:
-                        delayed_list.append(
-                            dask.delayed(_extract_holog_chunk)(
-                                dask.delayed(extract_holog_params)
+                        if parallel:
+                            delayed_list.append(
+                                dask.delayed(_extract_holog_chunk)(
+                                    dask.delayed(extract_holog_params)
+                                )
                             )
-                        )
+                        else:
+                            _extract_holog_chunk(extract_holog_params)
+                        count += 1
                     else:
-                        _extract_holog_chunk(extract_holog_params)
-                    count += 1
-                else:
-                    logger.warning(f'[{function_name}]: DDI ' + str(ddi) + ' has no holography data to extract.')
+                        logger.warning(f'[{function_name}]: DDI ' + str(ddi) + ' has no holography data to extract.')
 
-    spw_ctb.close()
-    pol_ctb.close()
-    obs_ctb.close()
+        spw_ctb.close()
+        pol_ctb.close()
+        obs_ctb.close()
 
-    if parallel:
-        dask.compute(delayed_list)    
+        if parallel:
+            dask.compute(delayed_list)    
 
-    if count > 0:
-        logger.info(f"[{function_name}]: Finished processing")
-        holog_dict = _load_holog_file(holog_file=extract_holog_params["holog_name"], dask_load=True, load_pnt_dict=False)
-        extract_holog_params['telescope_name'] = telescope_name
+        if count > 0:
+            logger.info(f"[{function_name}]: Finished processing")
+            holog_dict = _load_holog_file(holog_file=extract_holog_params["holog_name"], dask_load=True, load_pnt_dict=False)
+            extract_holog_params['telescope_name'] = telescope_name
         
-        holog_attr_file = "{name}/{ext}".format(name=extract_holog_params['holog_name'], ext=".holog_attr")
+            holog_attr_file = "{name}/{ext}".format(name=extract_holog_params['holog_name'], ext=".holog_attr")
 
-        meta_data = _create_holog_meta_data(
-            holog_file=extract_holog_params['holog_name'], 
-            holog_dict=holog_dict,
-            input_params=extract_holog_params.copy()
-        )
+            meta_data = _create_holog_meta_data(
+                holog_file=extract_holog_params['holog_name'], 
+                holog_dict=holog_dict,
+                input_params=extract_holog_params.copy()
+            )
 
-        _write_meta_data(holog_attr_file, meta_data)
+            _write_meta_data(holog_attr_file, meta_data)
 
-        holog_mds = AstrohackHologFile(extract_holog_params['holog_name'])
+            holog_mds = AstrohackHologFile(extract_holog_params['holog_name'])
+            holog_mds._open()
+        
+            return holog_mds
+        else:
+            logger.warning(f"[{function_name}]: No data to process")
+            return None
+
         holog_mds._open()
-        
+
         return holog_mds
-    else:
-        logger.warning(f"[{function_name}]: No data to process")
-        return None
-
-    holog_mds._open()
-
-    return holog_mds
+    
+    except Exception:
+        logger.error("There was an error, extiing.")
 
 def _check_extract_holog_params(function_name, extract_holog_params):
