+++ src/astrohack/holog.py
@@ -99,58 +99,62 @@ def holog(
 
     """
     
-    holog_params = locals()
+    try:
+        holog_params = locals()
 
-    logger = _get_astrohack_logger()
+        logger = _get_astrohack_logger()
 
-    function_name = inspect.stack()[CURRENT_FUNCTION].function
+        function_name = inspect.stack()[CURRENT_FUNCTION].function
 
-    ######### Parameter Checking #########
-    holog_params = _check_holog_params(function_name=function_name, holog_params=holog_params)
+        ######### Parameter Checking #########
+        holog_params = _check_holog_params(function_name=function_name, holog_params=holog_params)
     
-    _check_if_file_exists(holog_params['holog_name'])
-    _check_if_file_will_be_overwritten(holog_params['image_name'], holog_params['overwrite'])
+        _check_if_file_exists(holog_params['holog_name'])
+        _check_if_file_will_be_overwritten(holog_params['image_name'], holog_params['overwrite'])
 
-    json_data = "/".join((holog_params['holog_name'], ".holog_json"))
+        json_data = "/".join((holog_params['holog_name'], ".holog_json"))
 
-    with open(json_data, "r") as json_file:
-        holog_json = json.load(json_file)
+        with open(json_data, "r") as json_file:
+            holog_json = json.load(json_file)
         
-    meta_data = _read_meta_data(holog_params['holog_name'], 'holog', 'extract_holog')
+        meta_data = _read_meta_data(holog_params['holog_name'], 'holog', 'extract_holog')
 
-    if holog_params["cell_size"] is None:
-        cell_size = np.array([-meta_data["cell_size"], meta_data["cell_size"]])
-        holog_params["cell_size"] = cell_size
+        if holog_params["cell_size"] is None:
+            cell_size = np.array([-meta_data["cell_size"], meta_data["cell_size"]])
+            holog_params["cell_size"] = cell_size
     
-    if holog_params["grid_size"] is None:
-        n_pix = int(np.sqrt(meta_data["n_pix"]))
-        grid_size = np.array([n_pix, n_pix])
-        holog_params["grid_size"] = grid_size
+        if holog_params["grid_size"] is None:
+            n_pix = int(np.sqrt(meta_data["n_pix"]))
+            grid_size = np.array([n_pix, n_pix])
+            holog_params["grid_size"] = grid_size
 
-    logger.info(f'[{function_name}]: Cell size: {str(cell_size)}, Grid size {str(grid_size)}')
+        logger.info(f'[{function_name}]: Cell size: {str(cell_size)}, Grid size {str(grid_size)}')
     
-    json_data = {
+        json_data = {
             "cell_size": holog_params["cell_size"].tolist(),
             "grid_size": holog_params["grid_size"].tolist()
-    }
+        }
     
-    with open(".holog_diagnostic.json", "w") as out_file:
-        json.dump(json_data, out_file)
+        with open(".holog_diagnostic.json", "w") as out_file:
+            json.dump(json_data, out_file)
 
-    if _dask_general_compute(function_name, holog_json, _holog_chunk, holog_params, ['ant', 'ddi'], parallel=parallel):
+        if _dask_general_compute(function_name, holog_json, _holog_chunk, holog_params, ['ant', 'ddi'], parallel=parallel):
         
-        output_attr_file = "{name}/{ext}".format(name=holog_params['image_name'], ext=".image_attr")
-        _write_meta_data(output_attr_file, holog_params.copy())
+            output_attr_file = "{name}/{ext}".format(name=holog_params['image_name'], ext=".image_attr")
+            _write_meta_data(output_attr_file, holog_params.copy())
         
-        image_mds = AstrohackImageFile(holog_params['image_name'])
-        image_mds._open()
+            image_mds = AstrohackImageFile(holog_params['image_name'])
+            image_mds._open()
         
-        logger.info(f'[{function_name}]: Finished processing')
+            logger.info(f'[{function_name}]: Finished processing')
 
-        return image_mds
-    else:
-        logger.warning(f"[{function_name}]: No data to process")
-        return None
+            return image_mds
+        else:
+            logger.warning(f"[{function_name}]: No data to process")
+            return None
+    
+    except Exception:
+        logger.error("There was an error, exiting.")
 
 
 def _check_holog_params(function_name, holog_params):
